backend/FBHelper/requirements.txt
Flask==3.0.3
python-dotenv==1.0.0
flask-cors==4.0.1
openai==1.36.1
